# docs/swagger.yaml
basePath: /
definitions:
  models.EndVisitRequest:
    properties:
      location:
        $ref: '#/definitions/models.Location'
      timestamp:
        example: "2023-05-15T17:00:00Z"
        type: string
    type: object
  models.Location:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
  models.Schedule:
    properties:
      clientName:
        type: string
      id:
        type: integer
      location:
        type: string
      shiftEnd:
        type: string
      shiftStart:
        type: string
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      visitStatus:
        type: string
    type: object
  models.StartVisitRequest:
    properties:
      location:
        $ref: '#/definitions/models.Location'
      timestamp:
        example: "2023-05-15T09:00:00Z"
        type: string
    type: object
  models.Stats:
    properties:
      completedToday:
        type: integer
      missedSchedules:
        type: integer
      totalSchedules:
        type: integer
      upcomingToday:
        type: integer
    type: object
  models.Task:
    properties:
      description:
        type: string
      id:
        type: integer
      reason:
        type: string
      scheduleId:
        type: integer
      status:
        type: string
    type: object
  models.UpdateTaskRequest:
    properties:
      reason:
        example: reason for update
        type: string
      status:
        example: completed
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the backend server for the Caregiver Shift Tracker application.
  title: Mini EVV Logger API
  version: "1.0"
paths:
  /schedules:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all caregiver schedules. Use the optional 'date'
        query parameter with the value 'today' to filter for schedules on the current
        day.
      parameters:
      - description: Filter by date
        enum:
        - today
        example: today
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: A list of schedules
          schema:
            items:
              $ref: '#/definitions/models.Schedule'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all schedules
      tags:
      - schedules
  /schedules/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves full details for a single schedule, including its assigned
        care activities and tasks.
      parameters:
      - description: Schedule ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Full details of the schedule
          schema:
            $ref: '#/definitions/models.Schedule'
        "400":
          description: Invalid Schedule ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Schedule not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a single schedule by ID
      tags:
      - schedules
  /schedules/{id}/end:
    post:
      consumes:
      - application/json
      description: Logs the end time and geolocation for a specific schedule. This
        marks the completion of a caregiver's visit.
      parameters:
      - description: Schedule ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: End Visit Payload
        in: body
        name: visit
        required: true
        schema:
          $ref: '#/definitions/models.EndVisitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Visit ended successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or Schedule ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: End a visit
      tags:
      - schedules
  /schedules/{id}/start:
    post:
      consumes:
      - application/json
      description: Logs the start time and geolocation for a specific schedule. This
        marks the beginning of a caregiver's visit.
      parameters:
      - description: Schedule ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Start Visit Payload
        in: body
        name: visit
        required: true
        schema:
          $ref: '#/definitions/models.StartVisitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Visit started successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or Schedule ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start a visit
      tags:
      - schedules
  /stats:
    get:
      consumes:
      - application/json
      description: Retrieves dashboard statistics, including counts for total, missed,
        upcoming, and completed schedules for the current day.
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard statistics
          schema:
            $ref: '#/definitions/models.Stats'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get dashboard stats
      tags:
      - stats
  /tasks/{taskId}/update:
    post:
      consumes:
      - application/json
      description: Updates the status of a specific task. The status can be 'pending',
        'completed', or 'not_completed'. If the status is 'not_completed', a reason
        is required.
      parameters:
      - description: Task ID
        example: 1
        in: path
        name: taskId
        required: true
        type: integer
      - description: Update Task Payload
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Task updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or Task ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update task status
      tags:
      - tasks
swagger: "2.0"